sandbox/test2/requirements.txt
# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Async support
redis>=5.0.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# MCP Protocol
mcp>=1.0.0

# LLM Integration
anthropic>=0.18.0
openai>=1.10.0

# Agent utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Task queue
celery>=5.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
